Move more things into more places (server startup works)
