+++ SocialSignOn.mjs
@@ -1,7 +1,7 @@
 'use strict';
 
 import jwt from 'jsonwebtoken';
-import { SendRecvJSON, SendRecvFormEncoded } from './common.mjs';
+import { SendRecvJSON, SendRecvFormEncoded } from './lib/common.mjs';
 export { handleSSORedirect };
 
 async function handleSSORedirect(provider, authcode, config) {
