+++ lib/config.mjs
@@ -1,6 +1,6 @@
 'use strict'
 
-import { overwriteObj, readJson } from '../common.mjs';
+import { overwriteObj, readJson } from './common.mjs';
 
 export { bootstrapConfig };
