+++ routers/QueryAPIController.mjs
@@ -1,8 +1,8 @@
 'use strict';
 
 import { default as express } from 'express';
-import * as cmn from '../common.mjs';
-import * as wutil from '../webutils.mjs';
+import * as cmn from '../lib/common.mjs';
+import * as wutil from '../lib/webutils.mjs';
 import { validateDemoQueryRequest } from '../DemoQueryHandler.mjs';
 
 export { createQueryController };
