+++ routers/UserAPIController.mjs
@@ -2,7 +2,7 @@
 
 import { default as express } from 'express';
 import { User } from '../models/User.mjs';
-import * as wutil from '../webutils.mjs';
+import * as wutil from '../lib/webutils.mjs';
 import { UserSavedData } from '../models/UserSavedData.mjs';
 import { CookieNotFoundError, NoUserForSessionError, SessionExpiredError,
    SessionNotFoundError, SessionNotUsableError, UserDeletedError
